requirements.txt
# zuper stuff
PyGeometry-z6>=2.0
zuper-commons-z6>=6.1.5
compmake-z6>=6.0.8,<7
quickapp-z6>=6.0.5,<7
reprep-z6>=6.0.3,<7
zuper-typing-z6>=6.1.0,<7

# libraries
frozendict>=2.3.0
cytoolz>=0.11.2
#pytz==2021.1
numpy>=1.21.2
scipy>=1.7.1
matplotlib>=3.4.3
shapely>=1.8.0
PyYAML>=5.4.1
# imageio~=2.9.0 # candidate for removal?
setuptools>=57.4.0
cachetools>=5.0.0
networkx>=2.6.3

commonroad-io==2022.1

# homotopies
do-mpc>=4.2.0
oyaml~=1.0
casadi~=3.5.5
decorator~=5.1.0
cycler~=0.11.0
